[#7] moved advancedShadow modifier to it's on utilities class and cleaned up token files
