+++ udsnative/src/main/java/com/telus/udsnative/components/card/Card.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.unit.dp
 import com.telus.udsnative.ThemeResolver
 import com.telus.udsnative.components.card.models.CustomPadding
 import com.telus.udsnative.models.ComponentResolver
+import com.telus.udsnative.utility.advancedShadow
 
 
 /**
@@ -78,39 +79,7 @@ fun Card(
     }
 }
 
-fun Modifier.advancedShadow(
-    color: Color = Color.Black,
-    alpha: Float = .3f,
-    cornersRadius: Dp = 0.dp,
-    shadowBlurRadius: Dp = 0.dp,
-    offsetY: Dp = 0.dp,
-    offsetX: Dp = 0.dp
-) = drawBehind {
 
-    val shadowColor = color.copy(alpha = alpha).toArgb()
-    val transparentColor = color.copy(alpha = 0f).toArgb()
-
-    drawIntoCanvas {
-        val paint = Paint()
-        val frameworkPaint = paint.asFrameworkPaint()
-        frameworkPaint.color = transparentColor
-        frameworkPaint.setShadowLayer(
-            shadowBlurRadius.toPx(),
-            offsetX.toPx(),
-            offsetY.toPx(),
-            shadowColor
-        )
-        it.drawRoundRect(
-            0f,
-            0f,
-            this.size.width,
-            this.size.height,
-            cornersRadius.toPx(),
-            cornersRadius.toPx(),
-            paint
-        )
-    }
-}
 
 @Preview
 @Composable
